+++ MovieQuiz/Models/QuizQuestion.swift
@@ -1,0 +1,14 @@
+//
+//  QuizQuestion.swift
+//  MovieQuiz
+//
+//  Created by Anton Filipchuk on 15.07.2023.
+//
+
+import Foundation
+
+struct QuizQuestion {
+    let image: String
+    let test: String
+    let correctAnswer: Bool
+}
